BUG: Fixes 32 bit overflow issue in Require Minimum Size Features
