+++ src/Plugins/SimplnxCore/src/SimplnxCore/Filters/RequireMinimumSizeFeaturesFilter.cpp
@@ -1,5 +1,7 @@
 #include "RequireMinimumSizeFeaturesFilter.hpp"
 
+#include "SimplnxCore/Filters/Algorithms/RequireMinimumSizeFeatures.hpp"
+
 #include "simplnx/DataStructure/DataArray.hpp"
 #include "simplnx/DataStructure/Geometry/ImageGeom.hpp"
 #include "simplnx/Filter/Actions/DeleteDataAction.hpp"
@@ -8,7 +10,6 @@
 #include "simplnx/Parameters/DataGroupSelectionParameter.hpp"
 #include "simplnx/Parameters/GeometrySelectionParameter.hpp"
 #include "simplnx/Parameters/NumberParameter.hpp"
-#include "simplnx/Utilities/DataGroupUtilities.hpp"
 #include "simplnx/Utilities/FilterUtilities.hpp"
 #include "simplnx/Utilities/SIMPLConversion.hpp"
 
@@ -28,221 +29,6 @@ constexpr int32 k_BadMinAllowedFeatureSize = -5555;
 constexpr int32 k_BadNumCellsPath = -5556;
 constexpr int32 k_ParentlessPathError = -5557;
 
-void assign_badpoints(DataStructure& dataStructure, const DataPath& featureIdsPath, SizeVec3 dimensions, const NumCellsArrayType::store_type& numCellsStoreRef)
-{
-  FeatureIdsArrayType::store_type* featureIds = dataStructure.getDataAs<FeatureIdsArrayType>(featureIdsPath)->getDataStore();
-  usize totalPoints = featureIds->getNumberOfTuples();
-
-  std::array<int64_t, 3> dims = {
-      static_cast<int64>(dimensions[0]),
-      static_cast<int64>(dimensions[1]),
-      static_cast<int64>(dimensions[2]),
-  };
-
-  std::vector<int32_t> neighbors(totalPoints * featureIds->getNumberOfComponents(), -1);
-
-  int32 good = 1;
-  int32 current = 0;
-  int32 most = 0;
-  int64 neighpoint = 0;
-
-  std::array<int64_t, 6> neighpoints = {-dims[0] * dims[1], -dims[0], -1, 1, dims[0], dims[0] * dims[1]};
-
-  usize counter = 1;
-  int64 count = 0;
-  int64 kstride = 0;
-  int64 jstride = 0;
-  int32 featurename = 0;
-  int32 feature = 0;
-  int32 neighbor = 0;
-  std::vector<int32> n(numCellsStoreRef.getNumberOfTuples(), 0);
-
-  while(counter != 0)
-  {
-    counter = 0;
-    for(int64 k = 0; k < dims[2]; k++)
-    {
-      kstride = dims[0] * dims[1] * k;
-      for(int64 j = 0; j < dims[1]; j++)
-      {
-        jstride = dims[0] * j;
-        for(int64 i = 0; i < dims[0]; i++)
-        {
-          count = kstride + jstride + i;
-          featurename = featureIds->getValue(count);
-          if(featurename < 0)
-          {
-            counter++;
-            most = 0;
-            for(size_t l = 0; l < neighpoints.size(); l++)
-            {
-              good = 1;
-              neighpoint = count + neighpoints[l];
-              if(l == 0 && k == 0)
-              {
-                good = 0;
-              }
-              if(l == 5 && k == (dims[2] - 1))
-              {
-                good = 0;
-              }
-              if(l == 1 && j == 0)
-              {
-                good = 0;
-              }
-              if(l == 4 && j == (dims[1] - 1))
-              {
-                good = 0;
-              }
-              if(l == 2 && i == 0)
-              {
-                good = 0;
-              }
-              if(l == 3 && i == (dims[0] - 1))
-              {
-                good = 0;
-              }
-              if(good == 1)
-              {
-                feature = featureIds->getValue(neighpoint);
-                if(feature >= 0)
-                {
-                  n[feature]++;
-                  current = n[feature];
-                  if(current > most)
-                  {
-                    most = current;
-                    neighbors[count] = neighpoint;
-                  }
-                }
-              }
-            }
-            for(int32 l = 0; l < neighpoints.size(); l++)
-            {
-              good = 1;
-              neighpoint = count + neighpoints[l];
-              if(l == 0 && k == 0)
-              {
-                good = 0;
-              }
-              if(l == 5 && k == (dims[2] - 1))
-              {
-                good = 0;
-              }
-              if(l == 1 && j == 0)
-              {
-                good = 0;
-              }
-              if(l == 4 && j == (dims[1] - 1))
-              {
-                good = 0;
-              }
-              if(l == 2 && i == 0)
-              {
-                good = 0;
-              }
-              if(l == 3 && i == (dims[0] - 1))
-              {
-                good = 0;
-              }
-              if(good == 1)
-              {
-                feature = featureIds->getValue(neighpoint);
-                if(feature >= 0)
-                {
-                  n[feature] = 0;
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    DataPath attrMatPath = featureIdsPath.getParent();
-    auto* parentGroup = dataStructure.getDataAs<BaseGroup>(attrMatPath);
-    std::vector<std::string> voxelArrayNames;
-    for(const auto& [identifier, sharedChild] : *parentGroup)
-    {
-      if(std::dynamic_pointer_cast<IDataArray>(sharedChild))
-      {
-        voxelArrayNames.push_back(sharedChild->getName());
-      }
-    }
-
-    // TODO: This loop could be parallelized. Look at NeighborOrientationCorrelation filter
-    for(size_t j = 0; j < totalPoints; j++)
-    {
-      featurename = featureIds->getValue(j);
-      neighbor = neighbors[j];
-      if(neighbor >= 0)
-      {
-        if(featurename < 0 && featureIds->getValue(neighbor) >= 0)
-        {
-          for(auto& voxelArrayName : voxelArrayNames)
-          {
-            auto arrayPath = attrMatPath.createChildPath(voxelArrayName);
-            auto* arr = dataStructure.getDataAs<IDataArray>(arrayPath);
-            arr->copyTuple(neighbor, j);
-          }
-        }
-      }
-    }
-  }
-}
-
-// -----------------------------------------------------------------------------
-std::vector<bool> remove_smallfeatures(FeatureIdsArrayType::store_type& featureIdsStoreRef, const NumCellsArrayType::store_type& numCells, const PhasesArrayType::store_type* featurePhases,
-                                       int32_t phaseNumber, bool applyToSinglePhase, int64 minAllowedFeatureSize, Error& errorReturn)
-{
-  size_t totalPoints = featureIdsStoreRef.getNumberOfTuples();
-
-  bool good = false;
-  int32 gnum;
-
-  size_t totalFeatures = numCells.getNumberOfTuples();
-
-  std::vector<bool> activeObjects(totalFeatures, true);
-
-  for(size_t i = 1; i < totalFeatures; i++)
-  {
-    if(!applyToSinglePhase)
-    {
-      if(numCells.getValue(i) >= minAllowedFeatureSize)
-      {
-        good = true;
-      }
-      else
-      {
-        activeObjects[i] = false;
-      }
-    }
-    else
-    {
-      if(numCells.getValue(i) >= minAllowedFeatureSize || featurePhases->getValue(i) != phaseNumber)
-      {
-        good = true;
-      }
-      else
-      {
-        activeObjects[i] = false;
-      }
-    }
-  }
-  if(!good)
-  {
-    errorReturn = Error{-1, "The minimum size is larger than the largest Feature.  All Features would be removed"};
-    return activeObjects;
-  }
-  for(size_t i = 0; i < totalPoints; i++)
-  {
-    gnum = featureIdsStoreRef.getValue(i);
-    if(!activeObjects[gnum])
-    {
-      featureIdsStoreRef.setValue(i, -1);
-    }
-  }
-  return activeObjects;
-}
 } // namespace
 
 std::string RequireMinimumSizeFeaturesFilter::name() const
@@ -385,69 +171,17 @@ IFilter::PreflightResult RequireMinimumSizeFeaturesFilter::preflightImpl(const D
 Result<> RequireMinimumSizeFeaturesFilter::executeImpl(DataStructure& dataStructure, const Arguments& filterArgs, const PipelineFilter* pipelineNode, const MessageHandler& messageHandler,
                                                        const std::atomic_bool& shouldCancel, const ExecutionContext& executionContext) const
 {
-  auto featurePhasesPath = filterArgs.value<DataPath>(k_FeaturePhasesPath_Key);
-  auto featureIdsPath = filterArgs.value<DataPath>(k_FeatureIdsPath_Key);
-  auto imageGeomPath = filterArgs.value<DataPath>(k_ImageGeomPath_Key);
-  auto numCellsPath = filterArgs.value<DataPath>(k_NumCellsPath_Key);
-  auto applyToSinglePhase = filterArgs.value<bool>(k_ApplySinglePhase_Key);
-  auto minAllowedFeatureSize = filterArgs.value<int64>(k_MinAllowedFeaturesSize_Key);
-  auto phaseNumber = filterArgs.value<int64>(k_PhaseNumber_Key);
+  RequireMinimumSizeFeaturesInputValues inputValues;
 
-  PhasesArrayType::store_type* featurePhases = applyToSinglePhase ? dataStructure.getDataAs<PhasesArrayType>(featurePhasesPath)->getDataStore() : nullptr;
+  inputValues.ApplySinglePhase = filterArgs.value<BoolParameter::ValueType>(k_ApplySinglePhase_Key);
+  inputValues.FeatureIdsPath = filterArgs.value<ArraySelectionParameter::ValueType>(k_FeatureIdsPath_Key);
+  inputValues.FeaturePhasesPath = filterArgs.value<ArraySelectionParameter::ValueType>(k_FeaturePhasesPath_Key);
+  inputValues.InputImageGeometryPath = filterArgs.value<GeometrySelectionParameter::ValueType>(k_ImageGeomPath_Key);
+  inputValues.MinAllowedFeaturesSize = filterArgs.value<Int64Parameter::ValueType>(k_MinAllowedFeaturesSize_Key);
+  inputValues.NumCellsPath = filterArgs.value<ArraySelectionParameter::ValueType>(k_NumCellsPath_Key);
+  inputValues.PhaseNumber = filterArgs.value<Int64Parameter::ValueType>(k_PhaseNumber_Key);
 
-  FeatureIdsArrayType::store_type& featureIdsStoreRef = dataStructure.getDataAs<FeatureIdsArrayType>(featureIdsPath)->getDataStoreRef();
-
-  auto& numCellsArrayRef = dataStructure.getDataRefAs<NumCellsArrayType>(numCellsPath);
-  NumCellsArrayType::store_type& numCellsStoreRef = numCellsArrayRef.getDataStoreRef();
-
-  if(applyToSinglePhase && featurePhases != nullptr)
-  {
-    usize numFeatures = featurePhases->getNumberOfTuples();
-
-    bool unavailablePhase = true;
-    for(size_t i = 0; i < numFeatures; i++)
-    {
-      if(featurePhases->getValue(i) == phaseNumber)
-      {
-        unavailablePhase = false;
-        break;
-      }
-    }
-
-    if(unavailablePhase)
-    {
-      std::string ss = fmt::format("The phase number {} is not available in the supplied Feature phases array with path {}", phaseNumber, featurePhasesPath.toString());
-      return MakeErrorResult(-5555, ss);
-    }
-  }
-
-  Error errorReturn;
-  std::vector<bool> activeObjects = remove_smallfeatures(featureIdsStoreRef, numCellsStoreRef, featurePhases, phaseNumber, applyToSinglePhase, minAllowedFeatureSize, errorReturn);
-  if(errorReturn.code < 0)
-  {
-    return {nonstd::make_unexpected(std::vector<Error>{errorReturn})};
-  }
-
-  auto& imageGeom = dataStructure.getDataRefAs<ImageGeom>(imageGeomPath);
-  assign_badpoints(dataStructure, featureIdsPath, imageGeom.getDimensions(), numCellsStoreRef);
-
-  DataPath cellFeatureGroupPath = numCellsPath.getParent();
-  size_t currentFeatureCount = numCellsStoreRef.getNumberOfTuples();
-
-  int32 count = 0;
-  for(const auto& value : activeObjects)
-  {
-    if(value)
-    {
-      count++;
-    }
-  }
-  std::string message = fmt::format("Feature Count Changed: Previous: {} New: {}", currentFeatureCount, count);
-  messageHandler(nx::core::IFilter::Message{nx::core::IFilter::Message::Type::Info, message});
-
-  nx::core::RemoveInactiveObjects(dataStructure, cellFeatureGroupPath, activeObjects, featureIdsStoreRef, currentFeatureCount, messageHandler, shouldCancel);
-
-  return {};
+  return RequireMinimumSizeFeatures(dataStructure, messageHandler, shouldCancel, &inputValues)();
 }
 
 namespace
